requirements.txt
langchain==1.2.8
langgraph==1.0.7
langchain-community==0.4.1
langchain-openai==1.1.7
langchain-ollama
langgraph-checkpoint-sqlite==3.0.3
langchain-chroma==1.1.0
langchain-mcp-adapters
sentence-transformers
chromadb==1.4.1
slack_sdk==3.39.0
python-dotenv==1.2.1
arxiv==2.4.0
ipython
slack-bolt
requests
beautifulsoup4
pypdf==6.6.2
fastapi[standard]
gunicorn
uvicorn